+++ CHANGELOG.md
@@ -21,6 +21,8 @@ needed. This is equivalent to summarizing all activity on a feature branch versu
 
 ## Unreleased
 ### Added
+* validate_videos - Validation script which checks that the videos.csv file is valid. Performs basic validation after obtaining the ORCIDs from the .zenodo.json file to
+ensure that the contributors listed in the csv file are listed in the zenodo config file. This ensures that we give credit where it is due.
 * validate_bibfile - Validation script which checks that the bibliography file is valid (duplicate citation keys, syntax errors). Beyond general validity it enforces some knowlege-base specific requirements (doi, and note fields are required though in most contexts they are optional).
 * validate_basic - Basic validation script for csv files. The script is configured via a json file containing the following dictionary:
   * data_required_column_names - Columns that cannot contain empty entries.